+++ android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="ai.mlc.mlcchat">
+    package="com.modelbest.minicpm">
 
     <uses-permission android:name="android.permission.INTERNET" />
     
@@ -42,7 +42,7 @@
             android:name="libOpenCL-pixel.so"
             android:required="false" />
         <activity
-            android:name=".MainActivity"
+            android:name="com.modelbest.minicpm.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@android:style/Theme.Material.NoActionBar">
