+++ android/app/src/main/java/com/modelbest/minicpm/NavView.kt
@@ -1,6 +1,5 @@
-package ai.mlc.mlcchat
+package com.modelbest.minicpm
 
-import android.app.Application
 import android.app.Activity
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
