+++ android/app/src/main/java/com/modelbest/minicpm/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package ai.mlc.mlcchat.ui.theme
+package com.modelbest.minicpm.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
